Prevent subqueries from being marked as constant expressions
Subqueries were being marked as constant expressions by the
`ReduceExpressionsRule`. This is because subqueries are call expressions
with their own arguments. Call expressions are assumed to be reducible
if all of their arguments are reducible. The problem is that subqueries
weren't considering the subquery relation when determining if the call
was reducible.

While it might be possible for some subqueries to be constants and
reducible, that would only apply to things like a `LogicalValues` node
and it would be better if that were handled by other planner rules. This
change updates the `analyzeCall` method to mark all subqueries as
irreducible. Their arguments, if any, will still be reduced if possible,
but the subquery itself will never mark itself as a constant and never
be reducible.