Remove changes to the planner and catalog reader
These are the central classes where we enter and can customize behavior.
Unfortunately, these specific implementations don't give us enough
options to update behavior.

Instead of modifying them, just going to copy the implementations with
the copyright headers and make modifications where we need them with
alternative implementations when we need it.